+++ include/engine/graph/RenderEngineOptions.h
@@ -44,6 +44,7 @@ namespace engine::graph {
     void setMaximumRecursionDepth(int depth);
     void setMaximumThreads(int threads);
     void setQueueSize(int queueSize);
+    void setIntegrator(std::string integrator);
     void setSampler(std::string sampler);
     void setSamplesPerPixel(int samples);
     void setViewPlane(std::string viewPlane);
@@ -51,6 +52,7 @@ namespace engine::graph {
     std::optional<int> maximumRecursionDepth() const;
     std::optional<int> maximumThreads() const;
     std::optional<int> queueSize() const;
+    std::optional<std::string> integrator() const;
     std::optional<std::string> sampler() const;
     std::optional<int> samplesPerPixel() const;
     std::optional<std::string> viewPlane() const;
@@ -59,6 +61,7 @@ namespace engine::graph {
     std::optional<int> m_maximumRecursionDepth;
     std::optional<int> m_maximumThreads;
     std::optional<int> m_queueSize;
+    std::optional<std::string> m_integrator;
     std::optional<std::string> m_sampler;
     std::optional<int> m_samplesPerPixel;
     std::optional<std::string> m_viewPlane;